Added ability to nice backup, but stalls backup... disabled it.
For some reason BackupPC cannot get the file list
if the client (ssh->rsync) is niced. Makes no sense.

Left it in, so in the future I won't write the code
again just to see it fail.

In short, it's there, but disabled (0). Good enough
till a fix is found.