Fix bug: renice was causing stall. pre/post could as well.
Short answer: don't allow anything to output in the script.

Long: We were allowing renice as well as the pre/post programs
to output their stdout and stderr to BackupPC. This caused it
to think that was the rsync version. renice was just saying
it had changed the nice level, and BackupPC was waiting for
files.

And when the pre/post progam errored out, sent the output to
BackupPC, and the script went on; when rsync was run it waited
for the commands, but BackupPC had already sent them (in reply
to the errored pre/post command's error message).

Now pre/post has their output sent to /dev/null, and if a
pre/post fails, the script will output to stderr,
and then it will exit. This way the user will see in the
backup log "pre-run failed" instead of just wondering what happened.